PB-1295 : add drag&drop support for 3D viewer
